Temporarily revert to f32 to get interpolation and material coloring working

+++ server/src/generation.rs
@@ -57,7 +57,7 @@ impl Chunk {
 					let index = x << 8 | y << 4 | z;
 					let level_coordinates = chunk_origin_level_coordinates + vector![x as f32, y as f32, z as f32];
 					let distance = level_coordinates.metric_distance(&zero::<Vector3<_>>());
-					self.densities[index] = (256.0 * (level_radius - distance)) as u8;
+					self.densities[index] = 256.0 * (level_radius - distance);
 					self.materials[index] = material_map(distance);
 				}
 			}
